Modify the interaction between the MPI and the FHIR server
We split data between the MPI and the FHIR server to keep a single source of truth for clinical data (FHIR server) and patient data (MPI).

Instead of completely stripping out the patient resource when storing data to the FHIR server (as we do now) instead we create a gutted version of the patient. This keeps referential integrity in the FHIR server and enables operations that rely on there being a patient resource to be performed on the FHIR server (e.g. $everything and $summary).

This is a major change to how things worked in the past and might require a major version bump.