Modify the method for fetching all related patients
The gutted patient is now stored with the source id instead of the client registry interaction id. The gutted patient contains the link to the patient in the client registry. We now fetch the patient from hapi fhir and then use the interaction id in the link to fetch all the patient's links from the client registry