Fix bug in update logic
The patient's link (jempi interaction id) should not be changed. The update endpoint for jempi returns a different result from the registration endpoint. Patient Id returned for update is the golden id not the interaction id. The link was being updated with the golden id instead of the interaction id